+++ cvat-ui/src/reducers/plugins-reducer.ts
@@ -5,7 +5,6 @@
 
 import { PluginsActionTypes, PluginActions } from 'actions/plugins-actions';
 import { registerGitPlugin } from 'utils/git-utils';
-import { registerSAMPlugin } from 'utils/sam-utils';
 import { PluginComponent, PluginsState } from '.';
 
 const defaultState: PluginsState = {
@@ -73,8 +72,6 @@ export default function (state: PluginsState = defaultState, action: PluginActio
         }
         case PluginsActionTypes.GET_PLUGINS_SUCCESS: {
             const { list } = action.payload;
-
-            registerSAMPlugin();
 
             if (!state.list.GIT_INTEGRATION && list.GIT_INTEGRATION) {
                 registerGitPlugin();
